Set up a structure for importing timekeeping records from Mite.

+++ schema/20220117151719_create_table_timekeeping.sql
@@ -13,7 +13,7 @@ CREATE TABLE "timekeeping" (
   "mite_id" integer,
   "minutes" integer,
   "revenue" NUMERIC(10,2),
-  "hourly_rate" integer,
+  "hourly_rate" NUMERIC(10,2),
   CONSTRAINT "timekeeping_id_seq_pkey" PRIMARY KEY ("id")
 ) WITH (oids = false);
